# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pizza Delivery API",
        "description": "Документация REST API для проекта доставки пиццы. Здесь описаны доступные эндпоинты, форматы данных, авторизация и многое другое.",
        "contact": {
            "name": "Техподдержка Pizza Delivery",
            "email": "support@pizzadelivery.local"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1",
            "description": "Основной сервер (localhost)"
        }
    ],
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Получить список продуктов",
                "description": "Получить список всех продуктов.",
                "operationId": "1bfaa78d1c2c3848ab8165c5dadcad3e",
                "responses": {
                    "200": {
                        "description": "Список продуктов"
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Создать продукт",
                "description": "Создать новый продукт.",
                "operationId": "e8185e307706e2cd84bbcf5171e2f09d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Продукт создан"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/products/{product}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Получить продукт по ID",
                "description": "Получить один продукт по ID.",
                "operationId": "05f6ffcb49f47b4f4fc1a1c21ed60011",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "ID продукта",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Продукт найден"
                    },
                    "404": {
                        "description": "Продукт не найден"
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Обновить продукт",
                "description": "Обновить продукт.",
                "operationId": "c474d1318e5b8c3de1f4d52b6053111f",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "ID продукта",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Продукт обновлён"
                    },
                    "404": {
                        "description": "Продукт не найден"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Удалить продукт",
                "description": "Удалить продукт.",
                "operationId": "1a0a353d0842ff4f9347bb66ec53e3ea",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "ID продукта",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Продукт удалён"
                    },
                    "404": {
                        "description": "Продукт не найден"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}